+++ tests/test__init__.py
@@ -1,5 +1,3 @@
-# coding: utf-8
-
 import pytest
 
 from thermalprinter import ThermalPrinter
@@ -26,7 +24,7 @@ def test_default_values(port):
 
 
 def test_changing_good_values(port):
-    opt = {'heat_time': 120, 'heat_interval': 8, 'most_heated_point': 5}
+    opt = {"heat_time": 120, "heat_interval": 8, "most_heated_point": 5}
     printer = ThermalPrinter(port=port, baudrate=9600, **opt)
     assert printer._baudrate == 9600
     assert printer.heat_time == 120
@@ -35,7 +33,7 @@ def test_changing_good_values(port):
 
 
 def test_changing_bad_values(port):
-    opt = {'heat_time': 512, 'heat_interval': -42, 'most_heated_point': -42}
+    opt = {"heat_time": 512, "heat_interval": -42, "most_heated_point": -42}
     with pytest.raises(ThermalPrinterValueError):
         with ThermalPrinter(port=port, baudrate=9600, **opt):
             pass