+++ tests/test_chinese.py
@@ -1,6 +1,3 @@
-# coding: utf-8
-
-
 def test_default_value(printer):
     assert printer._chinese is False