+++ tests/test_image.py
@@ -1,10 +1,9 @@
-# coding: utf-8
-
 from os.path import abspath, dirname, realpath
 from pathlib import Path
 
 import pytest
 from PIL import Image
+
 from thermalprinter.exceptions import ThermalPrinterValueError
 
 
@@ -15,7 +14,7 @@ def test_empty_values(printer):
 
 def test_bad_type_str(printer):
     with pytest.raises(ThermalPrinterValueError):
-        printer.image('I am not an image ^^')
+        printer.image("I am not an image ^^")
 
 
 def test_bad_type_object(printer):
@@ -25,9 +24,9 @@ def test_bad_type_object(printer):
 
 def test_good(printer):
     cwd = dirname(realpath(abspath(__file__)))
-    printer.image(Image.open('{}/../gnu.png'.format(cwd)))
+    printer.image(Image.open("{}/../gnu.png".format(cwd)))
 
 
 def test_pathlib_path(printer):
-    image = Path(__file__).parent.parent / 'gnu.png'
+    image = Path(__file__).parent.parent / "gnu.png"
     printer.image(Image.open(image))